Cleanup. Added TileTool.cs, which is a base class for Tile-based tools.
