+++ Starforge.Vanilla/Tools/TileRectangleTool.cs
@@ -4,26 +4,15 @@ using Starforge.Editor;
 using Starforge.Editor.Tools;
 using Starforge.Map;
 using Starforge.Mod.API;
-using Starforge.Mod.Content;
 using Starforge.Vanilla.Actions;
 
 namespace Starforge.Vanilla.Tools {
     [ToolDefinition("TileRectangle")]
-    public class TileRectangleTool : Tool {
-        /// <remarks>The hint is set out of bounds (beyond upleft corner) so the hint does not appear when first selecting the tool.</remarks>
-        private Rectangle Hint = new Rectangle(-8, -8, 8, 8);
-
+    public class TileRectangleTool : TileTool {
         private Rectangle Hold = new Rectangle(-1, -1, 0, 0);
         private Point Start;
 
         public override string GetName() => "Tiles (Rectangle)";
-        public override bool CanSelectLayer() => true;
-
-        public override void Render() {
-            GFX.Draw.BorderedRectangle(Hint, Settings.ToolColor * 0.5f, Settings.ToolColor);
-            Hint.X = MapEditor.Instance.State.TilePointer.X * 8;
-            Hint.Y = MapEditor.Instance.State.TilePointer.Y * 8;
-        }
 
         public override void Update() {
             if (Input.Mouse.LeftClick) HandleClick();
@@ -77,10 +66,6 @@ namespace Starforge.Vanilla.Tools {
             Hold = new Rectangle(MapEditor.Instance.State.TilePointer.X, MapEditor.Instance.State.TilePointer.Y, 1, 1);
             Hint.Width = 8;
             Hint.Height = 8;
-        }
-
-        public override void RenderGUI() {
-            TileBrushTool.CreateGUITileTool();
         }
     }
 }
